+++ MovieTicketBooking.Web/Controllers/TheaterController.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.Operations;
 using MovieTicketBooking.Entities.Models;
 using MovieTicketBooking.Repositories.Interfaces;
 using MovieTicketBooking.Web.ViewModels.TheaterViewModels;
@@ -30,8 +31,6 @@ namespace MovieTicketBooking.Web.Controllers
                
             }
             ViewData["filterText"] = SearchText;
-           
-
 
             int totalItems = 0;
             if (!string.IsNullOrEmpty(SearchText))
@@ -72,14 +71,25 @@ namespace MovieTicketBooking.Web.Controllers
         [HttpPost]
         public async Task<IActionResult> AddTheater(AddTheaterViewModel vm)
         {
-            var theater = new Theater
+            if (ModelState.IsValid)
             {
-                TheaterName = vm.TheaterName,
-                Location = vm.Location,
-                Capacity = vm.Capacity,
-            };
-            await theaterRepo.Save(theater);
-            return RedirectToAction("TheaterList"); 
+                if (await theaterRepo.isTheaterExist(vm.TheaterName))
+                {
+                    TempData["error"] = "Theater already exists.";
+                    return View(vm);
+                }
+                var theater = new Theater
+                {
+                    TheaterName = vm.TheaterName,
+                    Location = vm.Location,
+                    Capacity = vm.Capacity,
+                };
+                await theaterRepo.Save(theater);
+                TempData["success"] = "Theater added successfully.";
+                return View(vm);
+            }
+            TempData["error"] = "Failed to add theater. Please check the input data.";
+            return View(vm);
         }
         [Authorize(Roles = "Admin")]
         public async Task<IActionResult> EditTheater(int id)
@@ -104,19 +114,30 @@ namespace MovieTicketBooking.Web.Controllers
         [HttpPost]
         public async Task<IActionResult> EditTheater(TheaterViewModel vm)
         {
-            var theater = new Theater
+            if (ModelState.IsValid)
             {
-                Id = vm.Id,
-                TheaterName = vm.TheaterName,
-                Location = vm.Location,
-                Capacity = vm.Capacity,
-            };
-           await theaterRepo.Edit(theater);
-            return RedirectToAction("TheaterList");
+                var theater = new Theater
+                {
+                    Id = vm.Id,
+                    TheaterName = vm.TheaterName,
+                    Location = vm.Location,
+                    Capacity = vm.Capacity,
+                };
+                await theaterRepo.Edit(theater);
+                TempData["success"] = "Theater updated successfully.";
+                return RedirectToAction("TheaterList");
+            }
+            TempData["error"] = "Failed to update theater. Please check the input data.";
+            return View(vm);
         }
         [Authorize(Roles = "Admin")]
         public async Task<IActionResult> RemoveTheater(int id)
         {
+            if (id ==null)
+            {
+                return NotFound();
+            }
+         
             var theater = await theaterRepo.GetById(id);
             TheaterViewModel vm = new TheaterViewModel
             {
@@ -125,6 +146,11 @@ namespace MovieTicketBooking.Web.Controllers
                 Location = theater.Location,
                 Capacity = theater.Capacity,
             };
+            bool hasAssociatedData = await theaterRepo.isMovieExist(vm.Id) || await theaterRepo.isTheaterWithShowtimesExist(vm.Id);
+            if (hasAssociatedData)
+            {
+                TempData["warning"] = "Associated movies or showtimes exist. Remove associated data before deleting the theater.";
+            }
             return View(vm);
         }
         [HttpPost]
@@ -137,7 +163,14 @@ namespace MovieTicketBooking.Web.Controllers
                 Location = vm.Location,
                 Capacity = vm.Capacity,
             };
+            bool hasAssociatedData = await theaterRepo.isMovieExist(vm.Id) || await theaterRepo.isTheaterWithShowtimesExist(vm.Id);
+            if (hasAssociatedData)
+            {
+                TempData["error"] = "Associated movies or showtimes exist. Remove associated data before deleting the theater.";
+                return View(vm);
+            }
             await theaterRepo.Remove(theater);
+            TempData["success"] = "Theater removed successfully!";
             return RedirectToAction("TheaterList");
         }
     }
